+++ app/Providers/AppServiceProvider.php
@@ -3,18 +3,22 @@
 namespace App\Providers;
 
 use App\Services\RagnarokSinks;
+use App\Services\Updater;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
 {
+    public $singletons = [
+        RagnarokSinks::class => RagnarokSinks::class,
+        Updater::class => Updater::class,
+    ];
+
     /**
      * Register any application services.
      */
     public function register(): void
     {
-        $this->app->singleton(RagnarokSinks::class, function () {
-            return new RagnarokSinks();
-        });
+        //
     }
 
     /**